Replace sVDBG and VDBG with mVerboseLoggingEnabled
Replace all instances of the private static fields sVDBG
and VDBG with the instance variable mVerboseLoggingEnabled.
This new name better reflects how this field is controlled
by the verbose debugging switch. This field is now an instance
variable (rather than a static variable) since it is set by the
enableVerboseDebugging() instance method of each class.

Also:
- Add a setter for mVerboseLoggingEnabled in WifiConfigStore, which
  is called from the same setter in WifiConfigManager.

BUG: 28321397
Change-Id: Icc3a2bf31b8d4090b89d429e1a1329c703116160
TEST: 'runtest frameworks-wifi' passes.