WifiConfigManager: Modify meaning of config.status
Change |config.status| field to track the status of a network within the
android framework rather than the status of network within
wpa_supplicant. |config.status| is enabled/disabled from only within
updateNetworkSelectionStatus now and will only change when
|enableNetwork| & |disableNetwork| public API's are invoked.

config.status is set to ENABLED when:
We set the status as enabled whenever network selection is set to
enabled.
config.status is set to DISABLED when:
(a) We set the status as disabled whenever network selection permanently
disables it. This is usually done when the app/user explicitly disables
the network or the current logged-in user changes. Newly created
networks are also DISABLED by default.
(b) We also need to mark all networks disabled when we start WPS.
(c) enableNetwork() public API with |disableOthers| flag set to true
disables all other saved networks.

While there,
1. Refactor the |updateNetworkStatus| method.

BUG: 28219928
TEST: Compiles & unit-test passes
TEST: Ran CTS tests
Change-Id: Id3cac801fa92e28984e451300f21f6f5b76de87a