Fix issue in WifiStateMachine.matchProviderWithCurrentNetwork().
This CL fixes a problem in matchProviderWithCurrentNetwork that prevented
the method from returning a deterministic value. This was caused by an
ordering issue related to setting the currently active Passpoint network
which is in turn used by matchProviderWithCurrentNetwork.

TEST: The functionality has been manually tested using the Hotspot 2.0 r2
app by adding instrumentation that forces a remediation (and a call to
matchProviderWithCurrentNetwork) right after association and correlating
log instrumentation with known values.

Bug=28600780

Change-Id: I798b87a61c335b87341e53e89f707ca9146fd0dd