[NAN] Treat GetCapabilities() as a regular command/response pair.
Original design treated GetCapabilities() as a notification which was
triggered whenever the HAL was initialized (even though under the hood
it was still a command/response). This can result in a race condition
if an "early" command needs the capabilities information. Converting to
a standard command/response will queue all other commands until the
capabilities are returned.

Specifically needed for NAN data-path which creates all interfaces (based
on capabilities) when it is first enabled.

Triggered whenever NAN is enabled - e.g. when Wi-Fi is enabled.

Bug: 26562610
Change-Id: I86375b318720b46cc92291484b7aa1d83928617b