WifiServiceImpl: handle verbose logging setting
Allow WifiServiceImpl to check the context for the verbose wifi logging
setting at startup.  This check is then used to set the verbose logging
variables in related objects.  The check was previously done in
WifiStateMachine.  When the wifi verbose logging is changed, this is
also now persisted with a call from WifiServiceImpl instead of
WifiStateMachine.

This CL additionally removes a method to get the verbose logging setting
from WifiConfigManager.  WifiStateMachine instead uses the variable it
already maintains.  All verbose logging statements in WifiStateMachine
are also now checked with mVerboseLoggingEnabled and the DBG variable is
removed.

Cleaned up unused imports in WifiStateMachine and WifiConfigManager.

BUG: 29009916
TEST: manually checked for verbose log messages after boot and after
changing verbose logging on and off
TEST: runtest frameworks-wifi

Change-Id: I0bb27bf1e4bf67f49a8976e357cb3f7d49d8ce66