WifiInjector: add WifiServiceImpl objects
Move objects created in WifiServiceImpl into WifiInjector to facilitate
unittests.  The first call should be to create an instance of
WifiInjector.  This call takes the Context as a parameter and is used to
initialize the instance.  This instance is also stored in the class and
subsequent calls should use the WifiInjector.getInstance() call.
Exceptions are thrown if the constructor is called multiple times or
will a null Context parameter.  An exception is also thrown if
getInstance is called before the WifiInjector constructor has been
called.

This change also allows the setWifiController method of
WifiLastResortWatchdog to be removed.

This change also removes the local variable for the
WifiLastResortWatchdog instance in WifiStateMachine.  Instead of holding
the LastResortWatchdog, calls from WifiStateMachine will get the
instance from WifiInjector when they are needed.  This is due to the
dependence triangle between WifiStateMachine, WifiController and
WifiLastResortWatchdog.

BUG: 28845563
Change-Id: Iad99cd9c03672b07d2d2eb67eb1f9821e402055f
TEST: runtest frameworks-wifi
TEST: manually checked connection to wifi, tethering and wifi reconnect