WifiConfigManagerNew: Change the config merge logic
Modify the merging logic from using |TextUtils.isEmpty()| to determine
if the field has a value to be merged or not to a simple null check. The
older check would prevent apps from resetting some fields later
if needed.

While there,
1. Add a unit test for |WifiEnterpriseConfig| object updates.
2. Add null checks for all publicly exposed fields in |WifiConfiguration|.

PS: |WifiEnterpriseConfig| copy constructor already merges the
provided configuration by checking if the value is set or not. So, we
don't need to add any special merging logic in WifiConfigManagerNew.

BUG: 29641570
Change-Id: Ic98147a9c07d2d3a878f103a9b2ea4988f96924f
TEST: Unit tests