Remove Venue Group and Venue Type parsing code
We currently do not use the Venue Group and Venue Type
codes that are parsed from the ANQP Venue Name Element,
so remove the code that parses and internally exposes these
fields.

Note: The Venue Name Duples in Venue Name ANQP element are
exposed on the ScanResult public API, so keep the parsing code
for that field.

BUG: 30291919
TEST: Framework compiles successfully.

Change-Id: I7c20daec9e5c9fe41253fb322c2cfce231846701