WifiLogger: rename to WifiDiagnostics
WifiLogger is a bit of a misnomer, as the module is
really about gathering diagnostics generally, and
not just logs in the strict sense.

Moreover, using "log" in the name might lead to
confusion with the logging code we're going to
add to the wifi service.

Rename WifiLogger, to minimize confusion.

BUG=30566052
TEST=wifitests/runtest.sh

Change-Id: I2c2e0a1c250e48f4ffcbf14b59b30ae0e3c6a68e