WifiConfigManagerNew: Migration from old store (Part 1)
In the new config store design, none of the other users private
networks should even be in memory because the other user's store files
are not accessible to the current user. However as we load networks
from the old store (all networks were in a single file) and migrate to
the new store, we might have a bunch of private networks from other
users in memory. We cannot write these private networks to the
respective store file until the corresponding user logs in to the
device after upgrade. So, move all such networks to the shared store file.
When the corresponding user logs in, their private networks will be sent
from the shared store file to the specific user store file and
eventually the shared store file will not contain any user private networks.

Changes in the CL:
1. Don't split the configurations to be sent for shared vs user store file inside
|WifiConfigStoreData|. This is now accepted as 2 separate lists in the
constructor.
2. Hide the members of |WifiConfigStoreData| and add getters to retrieve
the necessaty data.
3. |WifiConfigManagerNew| will send all private networks in memory not
visible to the current user along with any shared networks to the
shared store file.
4. Changes to the existing tests to account for #1 & #2 above.

BUG: 29337176
TEST: Ran existing tests. Will add unit tests for the specific scenario in
the next CL.

Change-Id: If0507bdd5df4296cfe94e0ff5492c4f6b354021d