WifiConfigManagerNew: handle User Switches
On user switch WifiConfigManager needs to
1. Switch the user store file in WifiConfigStore.
2. Check if the user is already unlocked.
3. If user is already unlocked, reload all the configurations
from the store files. This would ensure that the old user's
private networks are hidden.
4. If user is not unlocked, wait for the user to unlock to reload all
the configurations from the store files.
5. Write the store files again to write any private networks of the new
user that were in the shared store perviously to the user's store file.

Since we no longer have a single store file for all users, we don't need
the |ConfigurationMap.handleUserSwitch| method which was essentially
doing the same thing as above (but all user's networks were always
loaded in memory). So mark the method deprecated and remove it once the
old WifiConfigManager is removed.

BUG: 30783976
TEST: Added unit test

Change-Id: I4269b27ab5534b2e4ed4dfed1abd8327d354c66a