WifiConfigManagerNew: Migration from legacy stores
Integrated the new |WifiConfigStoreLegacy| into the new WifiConfigManager.
Every invocation of |loadFromStore| will check if there is any data to
be migrated from legacy store files if the new store files are not
present. If legacy store data is present, it will migrate all of the
data from those stores and delete them permanently.
The deletion of legacy stores is commented out in the first roll out of
the new config store changes. This will be re-enabled once things are
stable so that we can revert to this data if needed.

BUG: 29337176
Change-Id: I35d626ed8bae70c676cb10b307cb116ed837a529
TEST: Existing unit tests.