WSM: Handle external connection requests
The current implementation of enableNetwork() with disableOthers flag
set to true returns failure if the app does not hold the
OVERRIDE_WIFI_CONFIG permission. This is not backward compatible.

Instead, handle any external connect requests
(using enableNetwork, connect or save_network) using the following logic:
1. If the app holds the OVERRIDE_WIFI_CONFIG permission, immediately
trigger a connection to the requested network.
2, If not, force a connectivity manager scan and let QNS pick the best
network to connect to.

While there,
Remove the persist flag from QNS & WCM and rename the network
select method. The new name better reflects that we're just setting a
connect choice and not actually initiaing a connection.

BUG: 31272462
BUG: 31275240
BUG: 31216628
TEST: Unit tests
TEST: The WifiUtil.apk app used for smoke test verification now can call
enableNetwork successfully and ensure that we connect to the required
network.
TEST: android.net.wifi.cts.WifiManagerTest#testWifiManagerNetWork passes
now.

Change-Id: I540143e46dae55cd114711350b3676cd071f9930