WifiConfigStore: Ignore store dir creation failure
Modify |createFile| to always return a valid |StoreFile| instance even
if it fails to create the store directory. This may fail for user store
files if the user has not unlocked when wifi service is loaded for
example.
The user store file not being readable/writeable is already handled
specially in |StoreFile.read()| & |StoreFile.write()|. So, there is no
need to return a null here.

BUG: 31470077
Change-Id: Ibf9d1a6f8250548545200da43aac3d2ea7dfed65