WiFi Metrics: Add RSSI delta metric (poll v scan)
Created a new metric that logs the difference in RSSI values between a
ScanResult and the first RSSI poll received after attempting to connect
to the AP in that ScanResult.

RSSI poll must come within 3 seconds of framework's connections attempt,
and is not counted if the connection attempt fails.
The value logged is:
  delta = poll.rssi - scanResult.rssi

While I was here: Updated wifi.proto to match downstream placeholder
metrics.

Test: Added unit tests
Bug: 31239324
Merged-In: Ice3768df3e8b5c05ff74e7d75e205748403430dc
Change-Id: Ice3768df3e8b5c05ff74e7d75e205748403430dc