WifiConfigManager: Don't load shared networks on user unlock/switch
Currently, we reload all the store data (private + shared) on user
switching/first unlock. Loading of shared store data is unnecessary and
causes undesired side effect of the shared networks changing it's ID
after the reload. Instead, add a new method in WifiConfigStore to only
read the user's store file (holding private network configurations).

Changes in the CL:
1. Modify the |WifiConfigStore.switchUser| method to switch the user
store and also read out the data from the user store file.
2. When handling user switches/unlocks, delete only the private networks
of the old user from memory and then load the new user's private
networks.

TODO:  This CL will handle the case where the device was connected to a
shared network when the user unlocked or switched. Still need to initiate
disconnect if we're connected to one of the old user's private networks.
This will be added in a follow up CL.

Bug: 32802464
Test: Added Unit tests.
Test: Will send for regression testing.
Change-Id: I84f6f011bbd851f986e4294262af7a47e525524b