payment: Do not reset the default payment is non-null.
Initially the default payment service would be reset if the curently set
application is not available, or the settings would change the default
to another service, in case there was only one that was found.

Since this setting is now been backed up, it's possible that the application
is in process of getting installed during a restore. Cleanest solution is to
leave the default as is, if it is set. And it can be changed by the user
in case it needs to be different than the backed-up setting.

Bug: 32881476
Test: manual; Set default, remove application, default is unchanged

Change-Id: Iecdede121ead2a4e6d31697254cc70a8f1d7f9df
Signed-off-by: Ruchi Kandoi <kandoiruchi@google.com>