WifiStateMachine: remove service dependence
The constructor for WifiStateMachine currently depends on other wifi
related services starting in a particular order (ex, p2p starting before
wifi service).  This CL removes this dependence and delays retrieval of
other services until after the boot complete command.

Setting the local WifiApConfigStore variable is also moved to the
constructor for consistency with other locally stored state.

This CL additionally alters the WifiStateMachineTest to send a different
message to test log rec sizes when verbose logging settings are changed
(this set of tests seems out of place for WSM, but this CL is not meant
to address that problem).

Bug: 34137732
Test: frameworks/opt/net/wifi/tests/wifitests/runtests.sh
Test: frameworks/base/wifi/tests/runtests.sh
Test: manually tested wifi connection
Test: manually checked for Wifi Direct peers
Change-Id: I50fcc378bbbcc98ab833a01b9dd6c044c5f85e3d