WifiMonitor doesn't handle incorrect password authentication failure
This change updates dispatchEvent() in WifiMonitor, that parses native
events, to handle auth failure due to incorrect password. The absence
of this check causes the SupplicantTracker to send wrong authentication
status in the broadcast intent for Supplicant state change.

WifiMonitor also sends a message to subscribed interfaces with the
authentication failure event. In case of a wrong password, the event
SSID_TEMP_DISABLE is generated subsequently. WifiStateMachine, which
recieves both events, will erroneously update the WifiConfigManager
twice for the same authentication failure. To avoid this, this CL adds
a reason code in the message which can be used to conditionally update
WifiConfigManager.

Bug: 30080982
Test: Functionality tests and unit tests
Change-Id: I3154e6b52ace0b7c2295dc796325747439c0ac43