WifiStateMachine: Allow all apps to connect to a network
We had previously disallowed non settings apps from initiating
connection to a specific network for O using |disableOthers| flag in
WifiManager.enableNetwork.
But, we're relaxing this for O. We'll enforce this back in the next
release.

Note: We're adding an API surface with better semantics
(Network Specifier) for apps to make such requests in the future.

Bug: 34520151
Test: Unit tests
Change-Id: I78c382eab67a6dcff18627b6ca8ee282a04b3dde