WifiStateMachine: Prepare for HIDL integration
Changes in the CL:
1. As a precursor to plumbing through the WifiVendorHal method to
WifiNative, add a |isStaMode| flag to startHal in WifiNative.
WifiVendorHal.startVendorHal() uses this boolean to configure the chip
in the correct mode.
2. Add a new CMD_INITIALIZE message in WifiStateMachine. This will be
used to initialize the HAlDeviceManager module (via
WifiVendorHal.initialize()). This needs to be done
before the WifiVendorHal.startVendorHal() is invoked, so triggering this
when we get notification that the system services are ready.
3. Also added error recovery in WifiVendorHal.start(), any failures will
trigger a HAL stop.
4. Add a flag to enable the HIDL calls.

Bug: 33384303
Test: Compiles
Test: Added unit tests
Test: Devices is able to boot up and configure the chip in STA mode
(via HIDL using some local changes not yet ready to commit).
Test: Devices is able to boot up and connect to an access point (not
via HIDL, making sure nothing is broken if we check this CL without the
rest of HIDL changes).

Change-Id: I8350a2e120edec4941f0e070478fa7e313a01bb3