[HAL Device Manager] Stop Wi-Fi if (re)started
When getting notification that Wi-Fi service is started
we don't know if this is a re-start of the service or
potentially of the framework itself. If the framework is
restarted then it is possible that the state of the service
and framework are out-of-sync.

Therefore, on every service restart make sure to reset it by
stopping Wi-Fi.

Bug: 34474043
Test: unit tests
Change-Id: I5968598756cae3c1181659ef153b36cada7c2c16