Always honor config specified BSSID
When WifiNetworkSelector recommends a network for connection and the
WifiConfiguration of the network has a BSSID specified, we always
use that BSSID for connection attempt no matter firmware controls
roaming or not.

If the WifiConfiguration doesn't have a BSSID specified,
WifiConnectivityManager doesn't specify a BSSID for connection attempt
if firmware controls roaming.

If the BSSID of the scan result for the recommended network
configuration doesn't match the config specified BSSID, we drop the
connection or roaming attempt.

While there, made the indent style consistent in the unit test file.

Bug: 35641886
Test: runtests.sh and manual tests.
Change-Id: I888a65f2f9d4cb418f43775a8ecd1035c0cead28