SupplicantStaNetworkHal: Add FT flags
This is a port of the existing logic in WifiSupplicantControl. Need to
add the FT flags when adding the network to wpa_supplicant when the
device supports it (specified in device config.xml). Changed the input
arguments to SupplicantStaNetworkHal & SupplicantStaIfaceHal so that it
takes an instance of the Context and use WifiMonitor to trigger
callbacks instead of directly sending events using WifiStateMachine
handler.

Bug: 33383725
Test: Unit tests
Change-Id: Id266835ef5dc3038057cab5b8f8da3a327f39ce2