WifiStateMachine: allow config removal
Allow FORGET_NETWORK and REMOVE_NETWORK commands to be handled in the
DefaultState.  This means we can now remove configs in any state.

Proper handling for loaded config checks are handled by
WifiConfigManager.  A test was added to confirm calling removeNetwork
with a random int properly returns false when no network configs are
loaded.

Bug: 32809235
Test: frameworks/opt/net/wifi/tests/wifitests/runtests.sh
Test: frameworks/base/wifi/tests/runtests.sh
Test: All automated Sanity tests passed
Change-Id: I66dbef26530e3376f00537a3f1241bfd4f8a0d8b