Passpoint: Change ANQP response handling
Currently, the ANQP done notification from WifiMonitor is used
as a trigger to fetch anqp data in PasspointEventHandler.
In the HIDL interface, the callback itself will contain all the
necessary ANQP data. So, change the currently handling to prepare for
integration with HIDL interface.

Changes in the CL:
1. Move the ANQP data fetching to WifiMonitor away from
PasspointEventHandler.
2. Change the params of the ANQP done result to include a newly created
struct (AnqpResult) which would contain the bssid and the ANPQ results.
3. Add a new public method in WifiMonitor to send the notification out
from WifiMonitor, which will be used by the HIDL interface in the future.
4. Remove the redundant PasspointEventHandler.parsANQPLines().

Note: There are no unit tests for any of these changes because this CL
is just moving things around and most of it is going to be removed when
we integrate with HIDL.

Bug: 35393853
Test: Connects to passpoint networks.
Test: Will send for regression tests.
Change-Id: I7fdf8fbdba13d267eb986db3f8a5854c606bcd4c