Passpoint: Change ICON response handling
Currently, the ICON done notification from WifiMonitor is used
as a trigger to fetch icon data in PasspointEventHandler.
In the HIDL interface, the callback itself will contain all the
necessary icon data. So, change the currently handling to prepare for
integration with HIDL interface.

Changes in the CL:
1. Move the icon data fetching to WifiMonitor away from
PasspointEventHandler.
2. Change the params of the icon done event to include the icon data.
3. Add a new public method in WifiMonitor to send the notification out
from WifiMonitor, which will be used by the HIDL interface in the
future.

Note: There are no unit tests for any of these changes because this CL
is just moving things around and most of it is going to be removed when
we integrate with HIDL.

Bug: 35393853
Test: Connects to passpoint networks.
Test: Will send for regression tests.
Change-Id: I197180c8a8ec8673e5e8fa29ba8bb51b026d44fb