WifiSupplicantControl: Remove this class (Part 1)
Make WifiSupplicantControl instance part of WifiNative and
expose all the public methods of WifiSupplicantControl in WifiNative.
This is a prelude to removing this class altogether. All of this methods
have been exposed in SupplicantStaIface and WifiNative will use that
once we switch over to HIDL.

Note: Some of the existing methods in WifiSupplicantControl were just
wrappers over WifiNative methods and hence removed them altogether.

Bug: 35421721
Test: Existing unit tests.
Test: Will send for regression tests.
Change-Id: I1c90c8d13bec583ff4b2989c8d62b910c5e73e26