hotspot2: add config parser for legacy Passpoint configuration file
For Hotspot 2.0 Release 1, most of the configuration data are
already stored as part of WifiConfiguration across multiple files
in N and older.  There are number of fields contained in the
legacy Passpoint configuration file
(/data/misc/wifi/PerProviderSubscription.conf) that needed in order
to complete the configuration.  Mainly:
    - Friendly Name
    - Roaming Consortium OI
    - Realm
    - IMSI (for SIM credential)

So add the parsing support for the legacy Passpoint configuration
file to retrieve the missing configuration fields.

Refer to the javadoc in LegacyPasspointConfigParser for the format
of the configuration file.

Bug: 34206769
Test: frameworks/opt/net/wifi/tests/wifitests/runtests.sh
Change-Id: I1d1a96a2534e123fcd06e91eaa9fc18af842bea1