WifiConfigStoreLegacy: read legacy Passpoint configuration file
Read and load Passpoint configuration from the legacy Passpoint
configuration file (PerProviderSubscription.conf).  Copy the missing
configuration fields to WifiConfiguration, so that a WifiConfiguration
will contained all configuration data for Hotspot 2.0 Release 1
configuration.

The WifiConfiguration for Passpoint configuration will be stored
temporarily in the share store until its owner logs in.  When the
configuration owner logs in, the configuration will be converted
to PasspointConfiguration and added to PasspointManager.  Then the
shared WifiConfiguration will be removed to complete the migration.
This will be done in the follow-on CLs.

Bug: 34206769
Test: frameworks/opt/net/tests/wifitests/runtests.sh
Change-Id: I946984822599b46e7e79babdf84272290ecf2431