SupplicantStaNetworkHal: Fix network extra parsing/writing
The HIDL interface doesn't need the string to have the quotes. So,
refactor the helper functions so that it parses/creates an unquoted
string. The existing socket interface will continue to use the quoted
strings.

Also,
Don't attempt to read the wpa_supplicant.conf file in legacy store migration
if HIDL is enabled.

Bug: 35663125
Test: Unit tests
Test: Migration from N to O using HIDL interface.
Change-Id: I2ee9f61f6e9f393608982952986c5dd980c998f9