Create user store when migrating legacy configs
Currently, at the time when the migration starts, the user
store for the current user might not have been initialized.
This can happen if the user unlock before the config store
is ever read after boot.

To fix it, we will setup the user store for the current user
if it is not setup yet, before persisting the migrated data
to the config store.

While there, fix a bug when retrieving the password for
enterpise config from the legacy wpa_supplicant.conf file.
The password stored in that file is enclosed in double quotes,
and when we set the password in WifiEnterpriseConfig double
quotes will also be added.  This results in the password being
enclosded in duplicated double quotes.  So fix it by removing
the double quotes when retrieving the password from
wpa_supplicant.conf file.

Bug: 35675098
Test: frameworks/opt/net/tests/wifitests/runtests.sh
Test: manually updating a bullhead from N to O and verify
      user specific configurations are migrated

Change-Id: I565e9dcea8c45ba21595c18161cb976fbb472bd0