hotspot2: migrate legacy Passpoint configurations
When the legacy Passpoint configuration is read from the legacy
store, it will represented as WifiConfiguration and stored in the
share store.  When the owner of the legacy Passpoint configuration
logs in, the configuration will be converted to PasspointConfiguration
and added to PasspointManager, then persist to the user store in the
new format.

Bug: 34206769
Test: frameworks/opt/net/wifi/tests/wifitests/runtests.sh
Test: manually upgrade a bullhead containing Boingo and Global Reach
      Passpoint profiles from N to O, verify configurations are
      maintained and still able to connect to those APs

Change-Id: Idcd858326968463cea0cdc99452ba67b13d2ac77