util: add utility functions for retrieving Hotspot 2.0 specific IEs
Instead of using the parsed IEs maintained in NetworkDetail, the interested
party can use the newly added utility functions to retrieve the elements
that it is interested in.  This allows us to remove the NetworkDetail
dependencies from Passpoint code.

I think if we ever want to maintain a copy of parsed IEs, it should be
in ScanResult instead of NetworkDetail.

Bug: 35888100
Test: frameworks/opt/net/wifi/tests/wifitests/runtests.sh

Change-Id: I69849728c93d3e1f9b6c7bda1e5c21cfdb40689a