hotspot2: add support for WifiStateMachine#syncGetMatchingWifiConfig
This function will try to match a Passpoint AP to an installed Passpoint
provider.  Once a match is found, it will generate and return a
WifiConfiguration for the given Passpoint AP, which will contain
all necessary credentials for connecting to it.

Note that, the generated configuration is not added to WifiConfigManager.
The consumer of this WifiConfiguration (e.g. Settings app) can use
this WifiConfiguration to initiate a connection via WifiManager's API.

This operation is only supported in client mode, since PasspointManager
might issue ANQP requests when matching scan result to a provider.

Bug: 35888100
Test: frameworks/opt/net/wifi/tests/wifitests/runtests.sh
Test: install Global Reach Passpoint profile and verify the Global
      Reach Passpoint AP in the Settings UI is shown with
      "available via xxx" text.

Change-Id: I9bbd57819326d285367e89a342f21da79fc33555