WifiConfigurationXmlUtil: persist additional fields for legacy Passpoint configuration
When migrating legacy Passpoint configurations, the configuration will only
get migrated when its owner logs in.  In the meantime, we will
need persist all the required data in the share store temporarily.
So persist the additional required fields for Passpoint to the storage
to avoid any data loss in the case when migration took multiple
boots to complete.

Bug: 36099588
Test: frameworks/opt/net/wifi/tests/wifitests/runtests.sh
Test: manual test updating a device from N to O, verify all configs
      are migrated.

Change-Id: I040dfd9f207ed1a9b2e3ccd50eca50560d6e1444