WifiStateMachine: set supplicant log level after supplicant starts
When we start supplicant, we want to set its logging level to
reflect the logging level of WifiService. At present, we attempt
to set supplicant's logging level immediately after we
enableSupplicant().

This is problematic, because enableSupplicant() does not actually
block until supplicant is ready. Hence, the request to set supplicant's
logging level can be (and sometimes is) dropped.

Resolve this problem, by waiting for WifiMonitor.startMonitoring()
to return, before setting supplicant's logging level. (startMonitoring
is a blocking call.)

Note that, with this change, we'll no longer have any chance of
seeing debug-level log messages in the window from when supplicant
is started, until startMonitoring() returns. But I don't think such
messages are very important anyway.

Bug: 33462714
Test: manual (see below)

Manual test
-----------
1. flash DUT
2. Settings -> Developer options -> Enable Wi-Fi Verbose Logging -> On
3. reboot
4. adb shell logcat -d | grep ' D wpa_supplicant: ' | wc -l
   -> should be non-zero

Change-Id: Iaf9e3e8c047367ae55086d4d680664c6d954ebee