hotspot2: request ANQP element for OSU Providers
Ideally, we would only want to request OSU Providers ANQP element
when no provider is installed and request all elements when at least one
provider is installed.  However, this would result in more complexity
on the cache management and the code.

Since the ANQP results will cached, the traffic hits should be minimized.
This can be optimized in the future if needed to.

Bug: 36174048
Test: frameworks/opt/net/wifi/tests/wifitests/runtests.sh
Change-Id: Ie948caf4f37b2ff206271bd1d34d14fc2061ead6