WifiServiceImpl: check device encryption at boot
When a device is booting, it may need to be decrypted.  This extra step
will restart system services, including wifi.  If this restart happens
while we are creating interfaces or loading firmware, we can get odd
exceptions end up crashing.  To avoid this, we add a check to see if the
device needs to be decrypted.  If it does, we do not start wifi.

Added unittests to cover both decrypted and encrypted boot.  Also with
and without wifi enabled.

Bug: 35416881
Test: frameworks/opt/net/wifi/tests/wifitests/runtests.sh
Test: frameworks/base/wifi/tests/runtests.sh
Test: wifi integration tests
Change-Id: I8d471c7d0345611f90782dd3bc1b2261957eee73