WifiServiceImpl: add new methods to start softap
Add the implementations for two new methods to control softap mode.
These methods will be called by other methods in WifiServiceImpl and
also by ConnectivityService.  The calls are not intended to be called by
applications and will be protected by the NETWORK_STACK permission.

Note: this CL is adding the calls, they will be used in an upcoming CL.

Bug: 36540346
Test: frameworks/opt/net/wifi/tests/wifitests/runtests.sh
Test: frameworks/base/wifi/tests/runtests.sh

Change-Id: I6bc664a32c5df03db1dffb8db20330d9dc904113