LocalOnlyHotspotRequestInfo: track LOHS requests
Add a new class to keep track of apps Binder instances that request
LocalOnlyHotspot.  When the app dies, it will trigger a callback to
remove the request.  We will also unlinkToDeath when the app calls to
stop using LocalOnlyHotspot.  This object also contains a Messenger that
will allow WifiServiceImpl to communicate updates to WifiManager for the
requesting app.

Bug: 36732530
Test: frameworks/opt/net/wifi/tests/wifitests/runtests.sh

Change-Id: Ia366a3cc7339cdea94a137d62905b15b72cf05c0