tests: don't crash on Log.wtf() in eng builds
We'd like to be able to run the unit tests on eng builds,
and get meaningful results. Unfortunately, at the moment,
any test that excerises a call to Log.wtf() will cause
the tests to abort (not fail -- abort).

The abort happens because Log.wtf() aborts on eng builds
(by design). While that likely makes sense for normal
code, it gets in the way of testing.

Resolve this, by installing a custom TerribleFailureHandler
before running our tests.

Note that this does mean that our code behaves differently
under eng-test than eng-normal. I think we're okay with that,
since the eng-test behavior now more closely reflects the
user-normal and userdebug-normal behavior.

Bug: 36981745
Test: tests/wifitests/runtests.sh (on aosp_bullhead-eng)

Change-Id: Ieb3a3a98eacdf7131efd1efc6a966f68724e701b