[AWARE] Fix race condition of Aware state machine & HAL
The Wi-Fi Aware HAL was started before the Aware state machine
was ready. In case where the HAL was able to obtain a handle on
the NAN interface it would attempt to enable Aware through the
state machine - resulting in an NPE.

Added an explicit start method to the Aware HAL which is called
in order.

Bug: 37613363
Test: unit tests and integration (sl4a) test
Change-Id: I66698bd926ca5d4bdb19b3025ed2f26c61028d21