Fix P2P provision discovery config method
Fixes some more bugs introduced with HIDL. When using a PIN for
provision discovery, the target config method should be the opposite of
the role (KEYPAD or DISPLAY) that the sender is using.
This flips the role for the target config method when sending the
request, and sends the correct broadcast when completing the
provisioning on both devices.

Bug: 38002403
Test: Manually tested via Links app, CTS GO Negotiation Test (PIN),
updated unit test
Change-Id: Ic6f775030c6dd1476f56677fb79c895d59bb25b4