WifiDiagnostics: disable ring-buffers on |user| builds
We've seen ring-buffer events cause a number of problems
due to overload or deadlock. While the ring-buffers events
aren't necessarily the root cause of these problems, the
ring-buffer events _are_ a significant factor.

Mitigate risk for end-users, by disabling ring-buffers
on |user| builds.

Note: this CL differs from the automerge-parent CL in
oc-dev. The difference is that we do not remove the
ringBufferSizeIsSmallByDefault() test. The reason for
this difference is to make it easier to revert
ag/2116039, when the time comes (b/36811399).

(When we revert ag/2116039, we might want to allow
ringbuffers for user builds on DR. And if we do allow
ringbuffers on user builds, we'll want to have the
ringBufferSizeIsSmallByDefault test.)

Bug: 37756562
Test: tests/wifitests/runtests.sh
Test: manual test (see below)

Manual test: (on bullhead)
 1. install |user| build
 2. connect to GoogleGuest
 3. adb shell pgrep system_server
 4. adb shell dumpsys wifi | grep "ring-buffer ="
    -> expect no matching lines
 5. adb shell pgrep system_server
 6. PIDs from steps 3 and 5 should match
 7. install |userdebug| build
 8. connect to GoogleGuest
 9. adb shell pgrep system_server
10. adb shell dumpsys wifi | grep "ring-buffer ="
    -> expect at least one match
11. adb shell pgrep system_server
12. PIDs from steps 9 and 11 should match

Change-Id: Ib7cf1f9e9442ed3c3191d2b20b0812afbbf62930