WifiServiceImpl: implement startLOHS
When an application requests LOHS, check current state for incompatible
modes, outstanding requests and then register a request object for that
caller.  If this is the first caller, WifiServiceImpl will then trigger
WifiController to start softap mode.  When the softap mode switch
completes, callers will be notified of the outcome via the
LocalOnlyHotspotCallback api.

Bug: 31466854
Test: frameworks/opt/net/wifi/tests/wifitests/runtests.sh
Change-Id: If54756914f4daf5aab1663e6d747ead68d42fd1e