WifiBackupRestore: Tighten key checks in string parsing
When the wpa_supplicant.conf file received during restore is parsed, we
look for specific known keys which we use to identify the network type.
Instead of checking for lines starting with "<KEY>", check for lines
starting with "<KEY>=" which is the general format for all wpa_supplicant
params in the network block.

PS: This fixes the associated bug because the backup data contained an
unknown field called "eapRetryCount" which would have previously matched
out check to see if it's an EAP network.

Bug: 38248167
Test: Unit tests.
Change-Id: Ib4e158c8ed4cd10db694d389e5296e280363f9c9