WifiServiceImpl: act on interface ip mode changes
When Connectivity service notifies WifiService of interface ip mode
changes, we should use them to update our registered LOHS requestors.
When the LOHS is ready to use, we will be notified about the LOCAL_ONLY
mode and trigger the onStarted callback.  If tethering starts, we should
notify the requestors via the onFailed call with an INCOMPATIBLE_MODE
error.  In the unlikely case of an IP failure, we should also trigger
the onFailed callback with the GENERIC error.

This CL also adds tests and handling for cases where additional LOHS
requests are added after the hotspot is already up and running.

Bug: 31466854
Test: frameworks/opt/net/wifi/tests/wifitests/runtests.sh
Change-Id: I1c6a25eca7b23f161ea4403a980292e82968c4c1