[AWARE] Clean-up after session if fw indicates it no longer exists
When a session terminates and informs the host, it's state is
cleaned-up (erased). However, if the session is configured not
to inform the host (e.g. for power saving) then it's state still
exists. An attempt to reconfigure such a session will be passed to
the fw which will call back a failure with a specific 'invalid ID'
error code. In such a case clean up the state.

Bug: 62054202
Test: Discovery:test_ttl* pass, unit-tests pass.
Change-Id: I89d751cb5bee4e2cfb5519f3d3965922f5215eda